Feature/new api methods

+++ app/src/main/java/com/scp/adminforquiz/mvp/OneQuizPresenter.java
@@ -3,12 +3,14 @@ package com.scp.adminforquiz.mvp;
 import com.arellomobile.mvp.InjectViewState;
 import com.arellomobile.mvp.MvpPresenter;
 import com.scp.adminforquiz.Constants;
+import com.scp.adminforquiz.R;
 import com.scp.adminforquiz.api.ApiClient;
 import com.scp.adminforquiz.model.QuizConverter;
 import com.scp.adminforquiz.model.db.Quiz;
 import com.scp.adminforquiz.model.db.QuizTranslation;
 import com.scp.adminforquiz.model.db.QuizTranslationPhrase;
 import com.scp.adminforquiz.model.db.dao.QuizDao;
+import com.scp.adminforquiz.preference.MyPreferenceManager;
 
 import java.util.List;
 
@@ -21,6 +23,7 @@ import io.reactivex.functions.Function3;
 import io.reactivex.schedulers.Schedulers;
 import kotlin.Triple;
 import ru.terrakok.cicerone.Router;
+import timber.log.Timber;
 import toothpick.Toothpick;
 
 @InjectViewState
@@ -34,6 +37,11 @@ public class OneQuizPresenter extends MvpPresenter<OneQuizView> {
     ApiClient apiClient;
     @Inject
     QuizConverter quizConverter;
+    @Inject
+    MyPreferenceManager preferences;
+    private Long quizAuthorId;
+    private Long quizTranslationAuthorId;
+    private Long quizTranslationPhraseAuthorId;
 
     private Long quizId;
     private CompositeDisposable compositeDisposable = new CompositeDisposable();
@@ -50,6 +58,8 @@ public class OneQuizPresenter extends MvpPresenter<OneQuizView> {
     protected void onFirstViewAttach() {
         super.onFirstViewAttach();
         Toothpick.inject(this, Toothpick.openScope(Constants.APP_SCOPE));
+        Timber.d("User Id from pref : %s", preferences.getUserId());
+        Timber.d("isAdmin from PREF : %s", preferences.getIsAdmin());
         compositeDisposable.add(Flowable.combineLatest(
                 quizDao.getQuizByIdOrErrorWithUpdates(quizId),
                 quizDao.getQuizTranslationsByQuizIdWithUpdates(quizId),
@@ -59,7 +69,16 @@ public class OneQuizPresenter extends MvpPresenter<OneQuizView> {
                 .map(o -> quizDao.getQuizWithTranslationsAndPhrases(quizId))
                 .subscribeOn(Schedulers.io())
                 .observeOn(AndroidSchedulers.mainThread())
-                .subscribe(quiz -> getViewState().showQuiz(quiz)));
+                .subscribe(quiz -> {
+                    quizAuthorId = quiz.authorId;
+                    for (QuizTranslation quizTranslation : quiz.quizTranslations) {
+                        quizTranslationAuthorId = quizTranslation.authorId;
+                        for (QuizTranslationPhrase quizTranslationPhrase : quizTranslation.quizTranslationPhrases) {
+                            quizTranslationPhraseAuthorId = quizTranslationPhrase.authorId;
+                        }
+                    }
+                    getViewState().showQuiz(quiz);
+                }));
     }
 
     @Override
@@ -69,15 +88,17 @@ public class OneQuizPresenter extends MvpPresenter<OneQuizView> {
     }
 
     public void deleteQuiz() {
-        compositeDisposable.add(apiClient.deleteNwQuizById(quizId)
-                .map(aBoolean -> quizDao.deleteQuizById(quizId))
-                .subscribeOn(Schedulers.io())
-                .observeOn(AndroidSchedulers.mainThread())
-                .doOnSubscribe(disposable -> getViewState().showProgressBar(true))
-                .doOnEvent((integer, throwable) -> getViewState().showProgressBar(false))
-                .subscribe(integer -> router.backTo(Constants.ALL_QUIZ_SCREEN),
-                        error -> getViewState().showError(error.toString())
-                ));
+        if (quizAuthorId.equals(preferences.getUserId()) || preferences.getIsAdmin()) {
+            compositeDisposable.add(apiClient.deleteNwQuizById(quizId)
+                    .map(aBoolean -> quizDao.deleteQuizById(quizId))
+                    .subscribeOn(Schedulers.io())
+                    .observeOn(AndroidSchedulers.mainThread())
+                    .doOnSubscribe(disposable -> getViewState().showProgressBar(true))
+                    .doOnEvent((integer, throwable) -> getViewState().showProgressBar(false))
+                    .subscribe(integer -> router.backTo(Constants.ALL_QUIZ_SCREEN),
+                            error -> getViewState().showError(error.toString())
+                    ));
+        } else getViewState().showError(R.string.deleteToastText);
     }
 
     public void goToAddTranslationFragment() {
@@ -85,40 +106,45 @@ public class OneQuizPresenter extends MvpPresenter<OneQuizView> {
     }
 
     public void approveQuizById(Long quizId, boolean approve) {
-        compositeDisposable.add(apiClient.approveNwQuizById(quizId, approve)
-                .map(nwQuiz -> quizDao.insert(quizConverter.convert(nwQuiz)))
-                .subscribeOn(Schedulers.io())
-                .observeOn(AndroidSchedulers.mainThread())
-                .doOnSubscribe(disposable -> getViewState().showProgressBar(true))
-                .doOnEvent((aLong, throwable) -> getViewState().showProgressBar(false))
-                .subscribe(aLong -> {
-                        }, error -> getViewState().showError(error.toString())
-                ));
+        if (preferences.getIsAdmin()) {
+            compositeDisposable.add(apiClient.approveNwQuizById(quizId, approve)
+                    .map(nwQuiz -> quizDao.insert(quizConverter.convert(nwQuiz)))
+                    .subscribeOn(Schedulers.io())
+                    .observeOn(AndroidSchedulers.mainThread())
+                    .doOnSubscribe(disposable -> getViewState().showProgressBar(true))
+                    .doOnEvent((aLong, throwable) -> getViewState().showProgressBar(false))
+                    .subscribe(aLong -> {
+                            }, error -> getViewState().showError(error.toString())
+                    ));
+        } else getViewState().showError(R.string.toastApproveDialog);
     }
 
-//    public void approveTranslationById(Long translationId, boolean approve) {
-//        compositeDisposable.add(apiClient.approveNwQuizTranslationById(translationId, approve)
-//                .map(nwQuiz -> quizDao.insert(quizConverter.convert(nwQuiz)))
-//                .subscribeOn(Schedulers.io())
-//                .observeOn(AndroidSchedulers.mainThread())
-//                .doOnSubscribe(disposable -> getViewState().showProgressBar(true))
-//                .doOnEvent((aLong, throwable) -> getViewState().showProgressBar(false))
-//                .subscribe(aLong -> {
-//                        }, error -> getViewState().showError(error.toString())
-//                ));
-//    }  TODO
-
+    public void approveTranslationById(Long translationId, boolean approve) {
+        if (preferences.getIsAdmin()) {
+            compositeDisposable.add(apiClient.approveNwQuizTranslationById(translationId, approve)
+                    .map(nwQuizTranslation -> quizDao.insertQuizTranslation(quizConverter.convertTranslation(nwQuizTranslation, quizId)))
+                    .subscribeOn(Schedulers.io())
+                    .observeOn(AndroidSchedulers.mainThread())
+                    .doOnSubscribe(disposable -> getViewState().showProgressBar(true))
+                    .doOnEvent((aLong, throwable) -> getViewState().showProgressBar(false))
+                    .subscribe(aLong -> {
+                            }, error -> getViewState().showError(error.toString())
+                    ));
+        } else getViewState().showError(R.string.toastApproveDialog);
+    }
 
     public void deleteTranslationById(Long translationId) {
-        compositeDisposable.add(apiClient.deleteNwQuizTranslationById(translationId)
-                .map(aBoolean -> quizDao.deleteQuizTranslationById(translationId))
-                .subscribeOn(Schedulers.io())
-                .observeOn(AndroidSchedulers.mainThread())
-                .doOnSubscribe(disposable -> getViewState().showProgressBar(true))
-                .doOnEvent((integer, throwable) -> getViewState().showProgressBar(false))
-                .subscribe(integer -> {
-                        }, error -> getViewState().showError(error.toString())
-                ));
+        if (quizTranslationAuthorId.equals(preferences.getUserId()) || preferences.getIsAdmin()) {
+            compositeDisposable.add(apiClient.deleteNwQuizTranslationById(translationId)
+                    .map(aBoolean -> quizDao.deleteQuizTranslationById(translationId))
+                    .subscribeOn(Schedulers.io())
+                    .observeOn(AndroidSchedulers.mainThread())
+                    .doOnSubscribe(disposable -> getViewState().showProgressBar(true))
+                    .doOnEvent((integer, throwable) -> getViewState().showProgressBar(false))
+                    .subscribe(integer -> {
+                            }, error -> getViewState().showError(error.toString())
+                    ));
+        } else getViewState().showError(R.string.deleteToastText);
     }
 
     public void goToUpdateTranslationDescriptionFragment(Long translationId) {
@@ -126,32 +152,34 @@ public class OneQuizPresenter extends MvpPresenter<OneQuizView> {
     }
 
     public void deletePhraseById(Long phraseId) {
-        compositeDisposable.add(apiClient.deleteNwQuizTranslationPhraseById(phraseId)
-                .map(aBoolean -> quizDao.deleteQuizTranslationPhraseById(phraseId))
-                .subscribeOn(Schedulers.io())
-                .observeOn(AndroidSchedulers.mainThread())
-                .doOnSubscribe(disposable -> getViewState().showProgressBar(true))
-                .doOnEvent((integer, throwable) -> getViewState().showProgressBar(false))
-                .subscribe(integer -> {
-                        }, error -> getViewState().showError(error.toString())
-                ));
+        if (quizTranslationPhraseAuthorId.equals(preferences.getUserId()) || preferences.getIsAdmin()) {
+            compositeDisposable.add(apiClient.deleteNwQuizTranslationPhraseById(phraseId)
+                    .map(aBoolean -> quizDao.deleteQuizTranslationPhraseById(phraseId))
+                    .subscribeOn(Schedulers.io())
+                    .observeOn(AndroidSchedulers.mainThread())
+                    .doOnSubscribe(disposable -> getViewState().showProgressBar(true))
+                    .doOnEvent((integer, throwable) -> getViewState().showProgressBar(false))
+                    .subscribe(integer -> {
+                            }, error -> getViewState().showError(error.toString())
+                    ));
+        } else getViewState().showError(R.string.deleteToastText);
     }
 
     public void goToAddPhraseFragment(Long quizTranslationId) {
         router.navigateTo(Constants.ADD_PHRASE_SCREEN, quizTranslationId);
     }
 
-//    public void approvePhraseById(Long phraseId, boolean approve) {
-//        compositeDisposable.add(apiClient.approveNwQuizTranslationPhraseById(phraseId, approve)
-//                .map(nwQuiz -> quizDao.insert(quizConverter.convert(nwQuiz)))
-//                .subscribeOn(Schedulers.io())
-//                .observeOn(AndroidSchedulers.mainThread())
-//                .doOnSubscribe(disposable -> getViewState().showProgressBar(true))
-//                .doOnEvent((aLong, throwable) -> getViewState().showProgressBar(false))
-//                .subscribe(aLong -> {
-//                        }, error -> getViewState().showError(error.toString())
-//                ));
-//    }
-
+    public void approvePhraseById(Long phraseId, Long translationId, boolean approve) {
+        if (preferences.getIsAdmin()) {
+            compositeDisposable.add(apiClient.approveNwQuizTranslationPhraseById(phraseId, approve)
+                    .map(nwQuizTranslationPhrase -> quizDao.insertQuizTranslationPhrase(quizConverter.convertTranslationPhrase(nwQuizTranslationPhrase, translationId)))
+                    .subscribeOn(Schedulers.io())
+                    .observeOn(AndroidSchedulers.mainThread())
+                    .doOnSubscribe(disposable -> getViewState().showProgressBar(true))
+                    .doOnEvent((aLong, throwable) -> getViewState().showProgressBar(false))
+                    .subscribe(aLong -> {
+                            }, error -> getViewState().showError(error.toString())
+                    ));
+        } else getViewState().showError(R.string.toastApproveDialog);
+    }
 }
-
